Prevent duplication of name parameter
The old code used to have a problem when you started from the main
page and just submitted no name. You'd get two names: the empty one
and the generated one, which was turned into an array, and then
stringified into ARRAY and an address.

This commit fixes this by still setting the property when given, but
now we're only pushing it to the list of provided properties if it
hasn't been defined. In the situation mentioned above, the property
was already defined (but empty) and so the test returned false and the
property was added a second time.