Fix README; fix config
The tests were printing a warning because of the unreachable Face
Generator URL that the default config file was suggesting, so I
removed it. The URL is still added in the Dockerfile for those who
need it. The README was updated to explain the situation.

Updated the Debian packages in the README.

Removed the note about older Mojolicious versions from the README.

Simplified the morbo and hypnotoad instructions.

Changed the single quotes to double quotes in the Commands
initialisation because that breaks the translations test.